requirements.txt
asgiref==3.7.2
Django==4.2.7
django-auth-ldap==4.6.0
ldap3==2.9.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-ldap==3.4.3
sqlparse==0.4.4
unicode==2.9
